refactor: move the non-path test cases over to ipfs
this introduces duplication between the two sites but I think it's quite
tolerable and at least more favourable than living in lockstep.
